scripts/methods/cellpose3/requirements.txt
# Cellpose3 requirements for NucWorm benchmark
cellpose==3.1.1.2
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
tifffile>=2021.7.2
opencv-python-headless>=4.9.0.80
matplotlib>=3.4.0
tqdm>=4.62.0
pandas>=1.3.0
pyyaml>=5.4.0
